+++ docs/conf.py
@@ -18,11 +18,14 @@ sys.path.insert(0, os.path.abspath('../'))
 # -- Project information -----------------------------------------------------
 
 project = 'EBCIC'
-copyright = '2020-2021 National Institute of Advanced Industrial Science and Technology (AIST)'
-author = 'Kazukuni Kobara, AIST'
+copyright = (
+    '2020-2021 '
+    'National Institute of Advanced Industrial Science and Technology (AIST)'
+)
+author = 'Kazukuni KOBARA, AIST'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.1'
+release = '0.0.2'
 
 
 # -- General configuration ---------------------------------------------------
@@ -30,8 +33,7 @@ release = '0.0.1'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon'
-]
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']